Revert "refactor: extract ExpressionInput"
This reverts commit caf9cd338cb5148c8c7e273e837ea7a985cac4ae.
