+++ packages/core/client/src/schema-settings/VariableInput/index.ts
@@ -9,4 +9,3 @@
 
 export * from './hooks';
 export * from './VariableInput';
-export * from './ExpressionInput';
